simplified accelwater
I did a lot of work to get back to the same place. I removed a lot of
'magic' from accelwater and made it more general.  This had a side
effect of making the logic more transparent and simpler.  Github seems
to do a poor job of matching changes to the files.  It was less crazy
that this diff make it look like